fintools-client-backtests-skill  /backtests/backend/pyproject.toml
[project]
name = "am-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "akshare>=1.17.22",
    "a2a-sdk[http-server]>=0.3.24",
    "fastapi>=0.115.0",
    "httpx>=0.27.0",
    "numpy==2.1.3",
    "pandas>=2.3.1",
    "pydantic>=2.0.0",
    "pymysql>=1.1.1",
    "python-dateutil>=2.9.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "scipy>=1.16.0",
    "sqlalchemy>=2.0.41",
    "tqdm>=4.67.1",
    "tushare>=1.4.21",
    "uvicorn>=0.30.0",
]
